feat: remove Gemini support
BREAKING CHANGE:
Migrate your Gemini test to Hermine or use html-reporter@4.
