+++ README.md
@@ -1,8 +1,8 @@
 # html-reporter
 
-Plugin for [gemini](https://github.com/gemini-testing/gemini) and [hermione](https://github.com/gemini-testing/hermione) which is intended to aggregate the results of tests running into html report.
+Plugin for [hermione](https://github.com/gemini-testing/hermione) which is intended to aggregate the results of tests running into html report.
 
-You can read more about gemini plugins [here](https://github.com/gemini-testing/gemini/blob/master/doc/plugins.md) and about hermione plugins [here](https://github.com/gemini-testing/hermione#plugins).
+You can read more about hermione plugins [here](https://github.com/gemini-testing/hermione#plugins).
 
 ## Installation
 
@@ -18,7 +18,7 @@ Plugin has following configuration:
 
 * **enabled** (optional) `Boolean` – enable/disable the plugin; by default plugin is enabled
 * **path** (optional) `String` - path to directory for saving html report file; by
-default html report will be saved into `gemini-report/index.html` inside current work
+default html report will be saved into `hermione-report/index.html` inside current work
 directory.
 * **saveErrorDetails** (optional) `Boolean` – save/don't save error details to json-files (to error-details folder); `false` by default.
 
@@ -48,39 +48,9 @@ Use `html_reporter_` prefix for the environment variables and `--html-reporter-`
 
 For example you can override `path` option like so:
 ```bash
-$ html_reporter_path=custom/dir gemini test
-$ gemini test --html-reporter-path custom/dir
+$ html_reporter_path=custom/dir hermione test
+$ hermione test --html-reporter-path custom/dir
 ```
-
-### Gemini Usage
-
-Add plugin to your `gemini` config file:
-
-```js
-module.exports = {
-    // ...
-    system: {
-        plugins: {
-            'html-reporter/gemini': {
-                enabled: true,
-                path: 'my/gemini-reports',
-                defaultView: 'all',
-                baseHost: 'test.com',
-                errorPatterns: [
-                    'Parameter .* must be a string',
-                    {
-                        name: 'Cannot read property of undefined',
-                        pattern: 'Cannot read property .* of undefined'
-                    }
-                ]
-            }
-        }
-    },
-    //...
-}
-```
-
-### Hermione Usage
 
 Add plugin to your `hermione` config file:
