+++ lib/cli-commands/gui.js
@@ -4,10 +4,10 @@ const runGui = require('../gui');
 const Api = require('../gui/api');
 const {GUI: commandName} = require('./');
 
-module.exports = (program, pluginConfig, tool) => {
+module.exports = (program, pluginConfig, hermione) => {
     // must be executed here because it adds `gui` field in `gemini` and `hermione tool`,
     // which is available to other plugins and is an API for interacting with the current plugin
-    const guiApi = Api.create(tool);
+    const guiApi = Api.create(hermione);
 
     program
         .command(`${commandName} [paths...]`)
@@ -18,6 +18,6 @@ module.exports = (program, pluginConfig, tool) => {
         .option('-a, --auto-run', 'auto run immediately')
         .option('-O, --no-open', 'not to open a browser window after starting the server')
         .action((paths, options) => {
-            runGui({paths, tool, guiApi, configs: {options, program, pluginConfig}});
+            runGui({paths, hermione, guiApi, configs: {options, program, pluginConfig}});
         });
 };
