+++ lib/gui/api/index.js
@@ -3,12 +3,12 @@
 const ApiFacade = require('./facade');
 
 module.exports = class Api {
-    static create(tool) {
-        return new Api(tool);
+    static create(hermione) {
+        return new Api(hermione);
     }
 
-    constructor(tool) {
-        this._gui = tool.gui = ApiFacade.create();
+    constructor(hermione) {
+        this._gui = hermione.gui = ApiFacade.create();
     }
 
     initServer(server) {
