+++ lib/gui/app.js
@@ -4,7 +4,7 @@ const path = require('path');
 const _ = require('lodash');
 const looksSame = require('looks-same');
 const Promise = require('bluebird');
-const ToolRunnerFactory = require('./tool-runner-factory');
+const ToolRunner = require('./tool-runner');
 
 const lookSameAsync = (img1, img2, opts) => {
     return new Promise((resolve, reject) => {
@@ -15,14 +15,14 @@ const lookSameAsync = (img1, img2, opts) => {
 };
 
 module.exports = class App {
-    static create(paths, tool, configs) {
-        return new this(paths, tool, configs);
+    static create(paths, hermione, configs) {
+        return new this(paths, hermione, configs);
     }
 
-    constructor(paths, tool, configs) {
-        const {program, pluginConfig} = configs;
+    constructor(paths, hermione, configs) {
+        const {pluginConfig} = configs;
 
-        this._tool = ToolRunnerFactory.create(program.name(), paths, tool, configs);
+        this._toolRunner = ToolRunner.create(paths, hermione, configs);
         this._pluginConfig = pluginConfig;
 
         this._browserConfigs = [];
@@ -30,32 +30,32 @@ module.exports = class App {
     }
 
     initialize() {
-        return this._tool.initialize();
+        return this._toolRunner.initialize();
     }
 
     finalize() {
-        this._tool.finalize();
+        this._toolRunner.finalize();
     }
 
     run(tests) {
         return _.isEmpty(tests)
-            ? this._tool.run()
+            ? this._toolRunner.run()
             : this._runWithoutRetries(tests);
     }
 
     _runWithoutRetries(tests) {
         if (_.isEmpty(this._browserConfigs)) {
-            this._browserConfigs = _.map(this._tool.config.getBrowserIds(), (id) => this._tool.config.forBrowser(id));
+            this._browserConfigs = _.map(this._toolRunner.config.getBrowserIds(), (id) => this._toolRunner.config.forBrowser(id));
         }
 
         this._disableRetries();
 
-        return this._tool.run(tests)
+        return this._toolRunner.run(tests)
             .finally(() => this._restoreRetries());
     }
 
     updateReferenceImage(failedTests = []) {
-        return this._tool.updateReferenceImage(failedTests);
+        return this._toolRunner.updateReferenceImage(failedTests);
     }
 
     _resolveImgPath(imgPath) {
@@ -63,7 +63,7 @@ module.exports = class App {
     }
 
     _getCompareOpts() {
-        const {tolerance, antialiasingTolerance} = this._tool.config;
+        const {tolerance, antialiasingTolerance} = this._toolRunner.config;
 
         return {tolerance, antialiasingTolerance, stopOnFirstFail: true, shouldCluster: false};
     }
@@ -104,11 +104,11 @@ module.exports = class App {
     }
 
     addClient(connection) {
-        this._tool.addClient(connection);
+        this._toolRunner.addClient(connection);
     }
 
     get data() {
-        return this._tool.tree;
+        return this._toolRunner.tree;
     }
 
     _disableRetries() {
