+++ lib/gui/server.js
@@ -10,9 +10,9 @@ const {MAX_REQUEST_SIZE, KEEP_ALIVE_TIMEOUT, HEADERS_TIMEOUT} = require('./const
 const {IMAGES_PATH, ERROR_DETAILS_PATH} = require('../constants/paths');
 const {logger} = require('../server-utils');
 
-exports.start = ({paths, tool, guiApi, configs}) => {
+exports.start = ({paths, hermione, guiApi, configs}) => {
     const {options, pluginConfig} = configs;
-    const app = App.create(paths, tool, configs);
+    const app = App.create(paths, hermione, configs);
     const server = express();
 
     server.use(bodyParser.json({limit: MAX_REQUEST_SIZE}));
