--- lib/gui/tool-runner-factory/hermione/index.js
@@ -1,73 +1,0 @@
-'use strict';
-
-const _ = require('lodash');
-const BaseToolRunner = require('../base-tool-runner');
-const Runner = require('../runner');
-const subscribeOnToolEvents = require('./report-subscriber');
-const {formatTests} = require('../utils');
-const {formatId, getShortMD5, mkFullTitle} = require('./utils');
-
-module.exports = class HermioneRunner extends BaseToolRunner {
-    constructor(paths, tool, configs) {
-        super(paths, tool, configs);
-
-        this._tests = {};
-    }
-
-    run(tests = []) {
-        const {grep, set: sets, browser: browsers} = this._globalOpts;
-        const formattedTests = _.flatMap([].concat(tests), (test) => formatTests(test));
-
-        return Runner.create(this._toolName, this._collection, formattedTests)
-            .run((collection) => this._tool.run(collection, {grep, sets, browsers}));
-    }
-
-    _handleRunnableCollection() {
-        this._collection.eachTest((test, browserId) => {
-            if (test.disabled || test.silentSkip) {
-                return;
-            }
-
-            const testId = formatId(test.id(), browserId);
-            this._tests[testId] = _.extend(test, {browserId});
-
-            test.pending
-                ? this._reportBuilder.addSkipped(test)
-                : this._reportBuilder.addIdle(test);
-        });
-
-        this._fillTestsTree();
-    }
-
-    _subscribeOnEvents() {
-        subscribeOnToolEvents(this._tool, this._reportBuilder, this._eventSource, this._reportPath);
-    }
-
-    _prepareUpdateResult(test) {
-        const {browserId, attempt} = test;
-        const fullTitle = mkFullTitle(test);
-        const testId = formatId(getShortMD5(fullTitle), browserId);
-        const testResult = this._tests[testId];
-        const {sessionId, url} = test.metaInfo;
-        const assertViewResults = [];
-
-        const imagesInfo = test.imagesInfo.map((imageInfo) => {
-            const {stateName, actualImg} = imageInfo;
-            const path = this._tool.config.browsers[browserId].getScreenshotPath(testResult, stateName);
-            const refImg = {path, size: actualImg.size};
-
-            assertViewResults.push({stateName, refImg, currImg: actualImg});
-
-            return _.extend(imageInfo, {expectedImg: refImg});
-        });
-
-        return _.merge({}, testResult, {assertViewResults, imagesInfo, sessionId, attempt, meta: {url}, updated: true});
-    }
-
-    _emitUpdateReference({refImg}, state) {
-        this._tool.emit(
-            this._tool.events.UPDATE_REFERENCE,
-            {refImg, state}
-        );
-    }
-};
