--- lib/gui/tool-runner-factory/index.js
@@ -1,12 +1,0 @@
-'use strict';
-
-const toolAdapters = {
-    gemini: require('./gemini'),
-    hermione: require('./hermione')
-};
-
-module.exports = {
-    create: (toolName, paths, tool, configs) => {
-        return toolAdapters[toolName].create(paths, tool, configs);
-    }
-};
