--- lib/gui/tool-runner-factory/runner/specific-test-runner/gemini.js
@@ -1,14 +1,0 @@
-'use strict';
-
-const SpecificTestRunner = require('./');
-
-module.exports = class GeminiSpecificTestRunner extends SpecificTestRunner {
-    _filter() {
-        this._collection.disableAll();
-
-        this._tests.forEach(({suite, state, browserId}) => {
-            const suiteName = suite.path.join(' ');
-            this._collection.enable(suiteName, {state: state.name, browser: browserId});
-        });
-    }
-};
