+++ lib/gui/tool-runner/index.js
@@ -1,10 +1,14 @@
 'use strict';
 
-const path = require('path');
 const _ = require('lodash');
+const path = require('path');
 const chalk = require('chalk');
 const Promise = require('bluebird');
-const ReportBuilderFactory = require('../../report-builder-factory');
+
+const Runner = require('./runner');
+const subscribeOnToolEvents = require('./report-subscriber');
+const {formatTests, formatId, getShortMD5, mkFullTitle} = require('./utils');
+const ReportBuilder = require('../../report-builder');
 const EventSource = require('../event-source');
 const utils = require('../../server-utils');
 const {findTestResult} = require('./utils');
@@ -13,15 +17,13 @@ const reporterHelper = require('../../reporter-helpers');
 const {UPDATED} = require('../../constants/test-statuses');
 
 module.exports = class ToolRunner {
-    static create(paths, tool, configs) {
-        return new this(paths, tool, configs);
+    static create(paths, hermione, configs) {
+        return new this(paths, hermione, configs);
     }
 
-    constructor(paths, tool, {program: globalOpts, pluginConfig, options: guiOpts}) {
-        this._toolName = globalOpts.name();
-
+    constructor(paths, hermione, {program: globalOpts, pluginConfig, options: guiOpts}) {
         this._testFiles = [].concat(paths);
-        this._tool = tool;
+        this._hermione = hermione;
         this._tree = null;
         this._collection = null;
 
@@ -30,12 +32,14 @@ module.exports = class ToolRunner {
         this._reportPath = pluginConfig.path;
 
         this._eventSource = new EventSource();
-        this._reportBuilder = ReportBuilderFactory.create(this._toolName, tool, pluginConfig);
-        this._reportBuilder.setApiValues(tool.htmlReporter.values);
+        this._reportBuilder = ReportBuilder.create(hermione, pluginConfig);
+        this._reportBuilder.setApiValues(hermione.htmlReporter.values);
+
+        this._tests = {};
     }
 
     get config() {
-        return this._tool.config;
+        return this._hermione.config;
     }
 
     get tree() {
@@ -55,7 +59,7 @@ module.exports = class ToolRunner {
     _readTests() {
         const {grep, set: sets, browser: browsers} = this._globalOpts;
 
-        return this._tool.readTests(this._testFiles, {grep, sets, browsers});
+        return this._hermione.readTests(this._testFiles, {grep, sets, browsers});
     }
 
     finalize() {
@@ -121,6 +125,63 @@ module.exports = class ToolRunner {
             utils.logger.warn(chalk.yellow(`Nothing to reuse in ${this._reportPath}`));
             return {};
         }
+    }
+
+    run(tests = []) {
+        const {grep, set: sets, browser: browsers} = this._globalOpts;
+        const formattedTests = _.flatMap([].concat(tests), (test) => formatTests(test));
+
+        return Runner.create(this._collection, formattedTests)
+            .run((collection) => this._hermione.run(collection, {grep, sets, browsers}));
+    }
+
+    _handleRunnableCollection() {
+        this._collection.eachTest((test, browserId) => {
+            if (test.disabled || test.silentSkip) {
+                return;
+            }
+
+            const testId = formatId(test.id(), browserId);
+            this._tests[testId] = _.extend(test, {browserId});
+
+            test.pending
+                ? this._reportBuilder.addSkipped(test)
+                : this._reportBuilder.addIdle(test);
+        });
+
+        this._fillTestsTree();
+    }
+
+    _subscribeOnEvents() {
+        subscribeOnToolEvents(this._hermione, this._reportBuilder, this._eventSource, this._reportPath);
+    }
+
+    _prepareUpdateResult(test) {
+        const {browserId, attempt} = test;
+        const fullTitle = mkFullTitle(test);
+        const testId = formatId(getShortMD5(fullTitle), browserId);
+        const testResult = this._tests[testId];
+        const {sessionId, url} = test.metaInfo;
+        const assertViewResults = [];
+
+        const imagesInfo = test.imagesInfo.map((imageInfo) => {
+            const {stateName, actualImg} = imageInfo;
+            const path = this._hermione.config.browsers[browserId].getScreenshotPath(testResult, stateName);
+            const refImg = {path, size: actualImg.size};
+
+            assertViewResults.push({stateName, refImg, currImg: actualImg});
+
+            return _.extend(imageInfo, {expectedImg: refImg});
+        });
+
+        return _.merge({}, testResult, {assertViewResults, imagesInfo, sessionId, attempt, meta: {url}, updated: true});
+    }
+
+    _emitUpdateReference({refImg}, state) {
+        this._hermione.emit(
+            this._hermione.events.UPDATE_REFERENCE,
+            {refImg, state}
+        );
     }
 };
