+++ lib/gui/tool-runner/report-subscriber.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const clientEvents = require('../../constants/client-events');
-const {RUNNING} = require('../../../constants/test-statuses');
-const {getSuitePath} = require('../../../plugin-utils').getHermioneUtils();
-const utils = require('../utils');
-const createHermioneWorkers = require('../../../workers/create-hermione-workers');
+const clientEvents = require('../constants/client-events');
+const {RUNNING} = require('../../constants/test-statuses');
+const {getSuitePath} = require('../../plugin-utils').getHermioneUtils();
+const utils = require('./utils');
+const createWorkers = require('../../workers/create-workers');
 
 let workers;
 
@@ -24,7 +24,7 @@ module.exports = (hermione, reportBuilder, client, reportPath) => {
     }
 
     hermione.on(hermione.events.RUNNER_START, (runner) => {
-        workers = createHermioneWorkers(runner);
+        workers = createWorkers(runner);
     });
 
     hermione.on(hermione.events.SUITE_BEGIN, (suite) => {
