+++ lib/gui/tool-runner/runner/all-test-runner.js
