+++ lib/gui/tool-runner/runner/specific-test-runner.js
@@ -1,6 +1,7 @@
 'use strict';
 
-const Runner = require('../runner');
+const Runner = require('./runner');
+const {mkFullTitle} = require('../utils');
 
 module.exports = class SpecificTestRunner extends Runner {
     constructor(collection, tests) {
@@ -16,6 +17,10 @@ module.exports = class SpecificTestRunner extends Runner {
     }
 
     _filter() {
-        throw new Error('Method must be implemented in child classes');
+        this._collection.disableAll();
+
+        this._tests.forEach((test) => {
+            this._collection.enableTest(mkFullTitle(test), test.browserId);
+        });
     }
 };
