+++ lib/gui/tool-runner/utils.js
@@ -1,6 +1,8 @@
 'use strict';
 
 const _ = require('lodash');
+const crypto = require('crypto');
+
 const {findNode} = require('../../static/modules/utils');
 
 const formatTestHandler = (browser, test) => {
@@ -11,6 +13,17 @@ const formatTestHandler = (browser, test) => {
         state: {name},
         browserId: browser.name
     };
+};
+
+exports.formatId = (hash, browserId) => `${hash}/${browserId}`;
+
+exports.getShortMD5 = (str) => {
+    return crypto.createHash('md5').update(str, 'ascii').digest('hex').substr(0, 7);
+};
+
+exports.mkFullTitle = ({suite, state}) => {
+    // https://github.com/mochajs/mocha/blob/v2.4.5/lib/runnable.js#L165
+    return `${suite.path.join(' ')} ${state.name}`;
 };
 
 exports.formatTests = (test) => {
