+++ lib/plugin-adapter.js
@@ -3,19 +3,18 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
 const parseConfig = require('./config');
-const ReportBuilderFactory = require('./report-builder-factory');
+const ReportBuilder = require('./report-builder');
 const utils = require('./server-utils');
 const cliCommands = require('./cli-commands');
 const PluginApi = require('./plugin-api');
 
 module.exports = class PluginAdapter {
-    static create(tool, opts, toolName) {
-        return new this(tool, opts, toolName);
+    static create(hermione, opts) {
+        return new this(hermione, opts);
     }
 
-    constructor(tool, opts, toolName) {
-        this._tool = tool;
-        this._toolName = toolName;
+    constructor(hermione, opts) {
+        this._hermione = hermione;
         this._config = parseConfig(opts);
     }
 
@@ -24,15 +23,15 @@ module.exports = class PluginAdapter {
     }
 
     addApi() {
-        this._tool.htmlReporter = new PluginApi();
+        this._hermione.htmlReporter = new PluginApi();
 
         return this;
     }
 
     addCliCommands() {
         _.values(cliCommands).forEach((command) => {
-            this._tool.on(this._tool.events.CLI, (commander) => {
-                require(`./cli-commands/${command}`)(commander, this._config, this._tool);
+            this._hermione.on(this._hermione.events.CLI, (commander) => {
+                require(`./cli-commands/${command}`)(commander, this._config, this._hermione);
                 commander.prependListener(`command:${command}`, () => this._run = _.noop);
             });
         });
@@ -41,24 +40,24 @@ module.exports = class PluginAdapter {
     }
 
     init(prepareData, prepareImages) {
-        this._tool.on(this._tool.events.INIT, () => this._run(prepareData, prepareImages));
+        this._hermione.on(this._hermione.events.INIT, () => this._run(prepareData, prepareImages));
 
         return this;
     }
 
     _run(prepareData, prepareImages = _.noop) {
-        const reportBuilder = ReportBuilderFactory.create(this._toolName, this._tool, this._config);
+        const reportBuilder = ReportBuilder.create(this._hermione, this._config);
         const generateReport = Promise
             .all([
-                prepareData(this._tool, reportBuilder, this._config),
-                prepareImages(this._tool, reportBuilder, this._config)
+                prepareData(this._hermione, reportBuilder, this._config),
+                prepareImages(this._hermione, reportBuilder, this._config)
             ])
             .then(() => reportBuilder.save())
             .then(utils.logPathToHtmlReport)
             .catch(utils.logError);
 
-        const endRunnerEvent = this._tool.events.RUNNER_END || this._tool.events.END_RUNNER;
+        const endRunnerEvent = this._hermione.events.RUNNER_END || this._hermione.events.END_RUNNER;
 
-        this._tool.on(endRunnerEvent, () => generateReport);
+        this._hermione.on(endRunnerEvent, () => generateReport);
     }
 };
