+++ lib/report-builder.js
@@ -4,30 +4,31 @@ const path = require('path');
 const Promise = require('bluebird');
 const _ = require('lodash');
 const fs = require('fs-extra');
-const {IDLE, RUNNING, SUCCESS, FAIL, ERROR, SKIPPED, UPDATED} = require('../constants/test-statuses');
-const {hasImage, logger, prepareCommonJSData, shouldUpdateAttempt} = require('../server-utils');
-const {setStatusForBranch, hasFails, allSkipped, hasNoRefImageErrors} = require('../static/modules/utils');
+
+const TestAdapter = require('./test-adapter');
+const {IDLE, RUNNING, SUCCESS, FAIL, ERROR, SKIPPED, UPDATED} = require('./constants/test-statuses');
+const {hasImage, logError, prepareCommonJSData, shouldUpdateAttempt} = require('./server-utils');
+const {setStatusForBranch, hasFails, allSkipped, hasNoRefImageErrors} = require('./static/modules/utils');
 
 const NO_STATE = 'NO_STATE';
 
 module.exports = class ReportBuilder {
-    static create(tool, pluginConfig, TestAdapter) {
-        return new ReportBuilder(tool, pluginConfig, TestAdapter);
+    static create(hermione, pluginConfig) {
+        return new ReportBuilder(hermione, pluginConfig);
     }
 
-    constructor(tool, pluginConfig, TestAdapter) {
+    constructor(hermione, pluginConfig) {
         this._tree = {name: 'root'};
         this._skips = [];
-        this._tool = tool;
+        this._hermione = hermione;
         this._apiValues = {};
         this._pluginConfig = pluginConfig;
-        this._TestAdapter = TestAdapter;
     }
 
     format(result, status) {
-        return result instanceof this._TestAdapter
+        return result instanceof TestAdapter
             ? result
-            : this._TestAdapter.create(result, this._tool, status);
+            : TestAdapter.create(result, this._hermione, status);
     }
 
     addIdle(result) {
@@ -245,7 +246,7 @@ module.exports = class ReportBuilder {
         return this.saveDataFileAsync()
             .then(() => this._copyToReportDir(['index.html', 'report.min.js', 'report.min.css']))
             .then(() => this)
-            .catch((e) => logger.warn(e.message || e));
+            .catch(logError);
     }
 
     saveDataFileAsync() {
@@ -293,7 +294,7 @@ module.exports = class ReportBuilder {
 
     _copyToReportDir(files) {
         return Promise.map(files, (fileName) => {
-            const from = path.resolve(__dirname, '../static', fileName);
+            const from = path.resolve(__dirname, 'static', fileName);
             const to = path.join(this._pluginConfig.path, fileName);
 
             return fs.copy(from, to);
