--- lib/suite-adapter/gemini-suite-adapter.js
@@ -1,32 +1,0 @@
-'use strict';
-
-const SuiteAdapter = require('./suite-adapter');
-
-module.exports = class GeminiSuiteAdapter extends SuiteAdapter {
-    get skipComment() {
-        return this._wrapSkipComment(this._suite.skipComment);
-    }
-
-    getUrl(opts = {}) {
-        const browserConfig = this._config.forBrowser(opts.browserId);
-        const url = browserConfig.getAbsoluteUrl(this._suite.url);
-
-        return super._configureUrl(url, opts.baseHost);
-    }
-
-    get fullUrl() {
-        return this._suite.fullUrl;
-    }
-
-    get path() {
-        return this._suite.path;
-    }
-
-    get file() {
-        return this._suite.file;
-    }
-
-    get fullName() {
-        return this._suite.fullName;
-    }
-};
