+++ lib/test-adapter.js
@@ -2,17 +2,18 @@
 
 const Promise = require('bluebird');
 const _ = require('lodash');
-const path = require('path');
-
-const tmp = require('tmp');
-const TestAdapter = require('./test-adapter');
-const HermioneSuiteAdapter = require('../suite-adapter/hermione-suite-adapter.js');
-const {getSuitePath} = require('../plugin-utils').getHermioneUtils();
-const {SUCCESS, FAIL, ERROR} = require('../constants/test-statuses');
-const {ERROR_DETAILS_PATH} = require('../constants/paths');
-const utils = require('../server-utils');
 const fs = require('fs-extra');
+const path = require('path');
+const tmp = require('tmp');
 const crypto = require('crypto');
+
+const SuiteAdapter = require('./suite-adapter');
+const {getSuitePath} = require('./plugin-utils').getHermioneUtils();
+const {SUCCESS, FAIL, ERROR, UPDATED} = require('./constants/test-statuses');
+const {ERROR_DETAILS_PATH} = require('./constants/paths');
+const utils = require('./server-utils');
+
+const globalCacheAllImages = new Map();
 
 function createHash(buffer) {
     return crypto
@@ -24,14 +25,17 @@ function createHash(buffer) {
 const globalCacheDiffImages = new Map();
 const testsAttempts = new Map();
 
-module.exports = class HermioneTestResultAdapter extends TestAdapter {
-    constructor(testResult, tool, status) {
-        super(testResult, tool, status);
+module.exports = class TestAdapter {
+    static create(testResult = {}, hermione, status) {
+        return new this(testResult, hermione, status);
+    }
 
-        this._tool = tool;
-        this._errors = this._tool.errors;
-        this._suite = HermioneSuiteAdapter.create(this._testResult);
-        this._imagesSaver = this._tool.htmlReporter.imagesSaver;
+    constructor(testResult, hermione, status) {
+        this._testResult = testResult;
+        this._hermione = hermione;
+        this._errors = this._hermione.errors;
+        this._suite = SuiteAdapter.create(this._testResult);
+        this._imagesSaver = this._hermione.htmlReporter.imagesSaver;
         this._testId = `${this._testResult.fullTitle()}.${this._testResult.browserId}`;
         this._errorDetails = undefined;
 
@@ -40,6 +44,101 @@ module.exports = class HermioneTestResultAdapter extends TestAdapter {
         }
 
         this._attempt = testsAttempts.get(this._testId) || 0;
+    }
+
+    get suite() {
+        return this._suite;
+    }
+
+    get sessionId() {
+        return this._testResult.sessionId || 'unknown session id';
+    }
+
+    get browserId() {
+        return this._testResult.browserId;
+    }
+
+    get imagesInfo() {
+        return this._testResult.imagesInfo;
+    }
+
+    set imagesInfo(imagesInfo) {
+        this._testResult.imagesInfo = imagesInfo;
+    }
+
+    _getImgFromStorage(imgPath) {
+        return globalCacheAllImages.get(imgPath);
+    }
+
+    getImagesFor(status, stateName) {
+        const refImg = this.getRefImg(stateName);
+        const currImg = this.getCurrImg(stateName);
+        const errImg = this.getErrImg();
+
+        const refPath = utils.getReferencePath(this, stateName);
+        const currPath = utils.getCurrentPath(this, stateName);
+        const diffPath = utils.getDiffPath(this, stateName);
+
+        if (status === SUCCESS || status === UPDATED) {
+            return {expectedImg: {path: this._getImgFromStorage(refPath), size: refImg.size}};
+        }
+
+        if (status === FAIL) {
+            return {
+                expectedImg: {
+                    path: this._getImgFromStorage(refPath),
+                    size: refImg.size
+                },
+                actualImg: {
+                    path: this._getImgFromStorage(currPath),
+                    size: currImg.size
+                },
+                diffImg: {
+                    path: this._getImgFromStorage(diffPath),
+                    size: {
+                        width: _.max([_.get(refImg, 'size.width'), _.get(currImg, 'size.width')]),
+                        height: _.max([_.get(refImg, 'size.height'), _.get(currImg, 'size.height')])
+                    }
+                }
+            };
+        }
+
+        if (status === ERROR) {
+            return {
+                actualImg: {
+                    path: this.state ? this._getImgFromStorage(currPath) : '',
+                    size: currImg.size || errImg.size
+                }
+            };
+        }
+
+        return {};
+    }
+
+    async saveBase64Screenshot(reportPath) {
+        if (!this.screenshot.base64) {
+            utils.logger.warn('Cannot save screenshot on reject');
+
+            return Promise.resolve();
+        }
+
+        const currPath = utils.getCurrentPath(this);
+        const localPath = path.resolve(tmp.tmpdir, currPath);
+        await utils.makeDirFor(localPath);
+        await fs.writeFile(localPath, new Buffer(this.screenshot.base64, 'base64'), 'base64');
+
+        return this._saveImg(localPath, currPath, reportPath);
+    }
+
+    async _saveImg(localPath, destPath, reportDir) {
+        if (!localPath) {
+            return Promise.resolve();
+        }
+
+        const res = await this._imagesSaver.saveImg(localPath, {destPath, reportDir});
+
+        globalCacheAllImages.set(destPath, res || destPath);
+        return res;
     }
 
     get attempt() {
