+++ lib/workers/create-workers.js
@@ -1,7 +1,7 @@
 'use strict';
 
 module.exports = (runner) => {
-    const workerFilepath = require.resolve('./hermione-worker');
+    const workerFilepath = require.resolve('./worker');
 
     return runner.registerWorkers(workerFilepath, ['saveDiffTo']);
 };
