+++ lib/workers/worker.js
