+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "html-reporter",
   "version": "4.13.1",
-  "description": "Plugin for gemini and hermione which is intended to aggregate the results of tests running into html report",
+  "description": "Plugin for hermione which is intended to aggregate the results of tests running into html report",
   "scripts": {
     "coverage": "nyc npm run test-unit",
     "lint": "eslint .",
@@ -29,7 +29,6 @@
     "node": ">= 8"
   },
   "keywords": [
-    "gemini",
     "hermione",
     "plugin",
     "html-reporter"
