+++ test/unit/lib/gui/tool-runner/report-subsciber.js
@@ -1,14 +1,14 @@
 'use strict';
 
 const {EventEmitter} = require('events');
-const reportSubscriber = require('lib/gui/tool-runner-factory/hermione/report-subscriber');
-const ReportBuilder = require('lib/report-builder-factory/report-builder');
+const reportSubscriber = require('lib/gui/tool-runner/report-subscriber');
+const ReportBuilder = require('lib/report-builder');
 const clientEvents = require('lib/gui/constants/client-events');
 const {RUNNING} = require('lib/constants/test-statuses');
-const utils = require('lib/gui/tool-runner-factory/utils');
+const utils = require('lib/gui/tool-runner/utils');
 const {stubTool, stubConfig} = require('test/unit/utils');
 
-describe('lib/gui/tool-runner-factory/hermione/report-subscriber', () => {
+describe('lib/gui/tool-runner/hermione/report-subscriber', () => {
     const sandbox = sinon.createSandbox();
     let reportBuilder;
     let client;
